+++ backend/src/main/java/com/mcmasterbaja/exceptions/StorageException.java
@@ -1,0 +1,12 @@
+package com.mcmasterbaja.exceptions;
+
+public class StorageException extends RuntimeException {
+
+  public StorageException(String message) {
+    super(message);
+  }
+
+  public StorageException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
