--- backend/src/main/java/com/mcmasterbaja/storage/exceptions/FileNotFoundException.java
@@ -1,13 +1,0 @@
-package com.mcmasterbaja.storage.exceptions;
-
-public class FileNotFoundException extends StorageException {
-
-    public FileNotFoundException(String message) {
-        super(message);
-    }
-
-    public FileNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    
-}
