--- backend/src/main/java/com/mcmasterbaja/storage/exceptions/MalformedCsvException.java
@@ -1,22 +1,0 @@
-package com.mcmasterbaja.storage.exceptions;
-
-import lombok.Getter;
-
-@Getter
-public class MalformedCsvException extends StorageException {
-    private String file;
-
-    public MalformedCsvException(String message) {
-        super(message);
-    }
-    
-    public MalformedCsvException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public MalformedCsvException(String message, String file, Throwable cause) {
-        super(message, cause);
-        this.file = file;
-    }
-    
-}
