--- backend/src/main/java/com/mcmasterbaja/storage/exceptions/StorageException.java
@@ -1,13 +1,0 @@
-package com.mcmasterbaja.storage.exceptions;
-
-public class StorageException extends RuntimeException {
-
-    public StorageException(String message) {
-        super(message);
-    }
-
-    public StorageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    
-}
